Extends the RescopeGlobalSymbols compiler pass to make all extern symbols reference the global scope explicitly. The change is meant to be 100% backward compatible.
The current implementation is browser dependend. As we are most likely
the only customer of this compiler pass, I think that is OK. It could,
of course, be changed to make "window" be configurable.

This change allows injecting the global scope into code payloads. We
use this to enable JS payloads to be used across friendly iframes;
in particular all +1 buttons on a page, reuse the same JS.

R=nicksantos,johnlenz
DELTA=69  (69 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4381


git-svn-id: http://closure-compiler.googlecode.com/svn/trunk@1839 b0f006be-c8cd-11de-a2e8-8d36a3108c74